requirements.txt
torch
torchvision
transformers
peft
safetensors
tensorboard
pyyaml
pandas
numpy
scikit-learn
tqdm
Pillow
decord
